+++ src/gdal_driver.cpp
@@ -231,6 +231,11 @@ GDAL_ASYNCABLE_DEFINE(Driver::create) {
 }
 
 /**
+ * @typedef CreateOptions
+ * @property {ProgressCb} [progress_cb]
+ */
+
+/**
  * Create a copy of a dataset.
  *
  * @throws Error
@@ -238,8 +243,9 @@ GDAL_ASYNCABLE_DEFINE(Driver::create) {
  * @param {string} filename
  * @param {gdal.Dataset} src
  * @param {string[]|object} [options=null] An array or object containing driver-specific dataset creation options
- * @param {ProgressCb} [progress_cb=undefined] {{{progress_cb}}}
  * @param {boolean} [strict=false] strict mode
+ * @param {CreateOptions} [jsoptions] additional options
+ * @param {ProgressCb} [jsoptions.progress_cb] {{{progress_cb}}}
  * @return {gdal.Dataset}
  */
 
@@ -251,8 +257,9 @@ GDAL_ASYNCABLE_DEFINE(Driver::create) {
  * @param {string} filename
  * @param {gdal.Dataset} src
  * @param {string[]|object} [options=null] An array or object containing driver-specific dataset creation options
- * @param {ProgressCb} [progress_cb=undefined] {{{progress_cb}}}. When specified in Promise mode, strict must also be present or progress_cb will be interpreted as a result callback.
  * @param {boolean} [strict=false] strict mode
+ * @param {CreateOptions} [jsoptions] additional options
+ * @param {ProgressCb} [jsoptions.progress_cb] {{{progress_cb}}}
  * @param {callback<gdal.Dataset>} [callback=undefined] {{{cb}}}
  * @return {Promise<gdal.Dataset>}
  */
@@ -289,10 +296,12 @@ GDAL_ASYNCABLE_DEFINE(Driver::createCopy) {
   }
 
   bool strict = false;
-  NODE_ARG_BOOL_OPT(4, "strict", strict);
+  NODE_ARG_BOOL_OPT(3, "strict", strict);
 
+  Local<Object> jsoptions;
   Nan::Callback *progress_cb = nullptr;
-  NODE_ARG_CB_OPT(3, "progress_cb", progress_cb);
+  NODE_ARG_OBJECT_OPT(4, "jsoptions", jsoptions);
+  if (!jsoptions.IsEmpty()) NODE_CB_FROM_OBJ_OPT(jsoptions, "progress_cb", progress_cb);
 
   GDALDriver *raw = driver->getGDALDriver();
   GDALDataset *raw_ds = src_dataset->get();